Factorizing abstract summarizer class
