+++ cli/bin/simple.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
-java -cp "target/classes:target/*" \
+java -Xmx6g -cp "target/classes:target/*" \
 edu.stanford.futuredata.macrobase.runner.SimpleRunner "$@"
